Manage listing page

+++ app/Http/Controllers/ListingController.php
@@ -61,6 +61,11 @@ class ListingController extends Controller
 
     // Update Listing Data
     public function update(Request $request, Listing $listing){
+        //ensure if logged in user is the owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -82,6 +87,10 @@ class ListingController extends Controller
 
      // Delete Listing
     public function destroy(Listing $listing) {
+        //ensure other users to not delete posts
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
